test: isolate MThreads FP8 MM regression coverage
